requirements.txt
albumentations==1.1.0
certifi==2021.10.8
imageio==2.13.3
joblib==1.1.0
minio==7.1.2
networkx==2.6.3
numpy==1.21.4
opencv-python-headless==4.5.4.60
packaging==21.3
Pillow==8.4.0
pyparsing==3.0.6
PyWavelets==1.2.0
PyYAML==6.0
qudida==0.0.4
scikit-image==0.19.0
scikit-learn==1.0.1
scipy==1.7.3
threadpoolctl==3.0.0
tifffile==2021.11.2
typing_extensions==4.0.1
urllib3==1.26.7
